move EditableGrid related models to /internal/editable dir
