+++ packages/components/src/internal/actions.ts
@@ -40,13 +40,11 @@ import {
     CellMessage,
     CellMessages,
     CellValues,
-    createGridModelId,
     EditorModel,
     EditorModelProps,
     IGridResponse,
     ValueDescriptor,
-    VisualizationConfigModel,
-} from './models';
+} from './components/editable/models';
 import { DataViewInfo } from './DataViewInfo';
 import { EditableColumnMetadata } from './components/editable/EditableGrid';
 
@@ -64,6 +62,7 @@ import { buildURL } from './url/AppURL';
 
 import { ViewInfo } from './ViewInfo';
 import { decimalDifference, genCellKey, getSortedCellKeys, parseCellKey } from './utils';
+import { createGridModelId } from './models';
 
 const EMPTY_ROW = Map<string, any>();
 let ID_COUNTER = 0;
@@ -658,21 +657,6 @@ export function getSelectedData(
                 reject(resolveErrorMessage(reason));
             })
     );
-}
-
-export function getVisualizationConfig(reportId: string): Promise<VisualizationConfigModel> {
-    return new Promise((resolve, reject) => {
-        Query.Visualization.get({
-            reportId,
-            name: undefined,
-            schemaName: undefined,
-            queryName: undefined,
-            success: response => {
-                resolve(VisualizationConfigModel.create(response.visualizationConfig));
-            },
-            failure: reject,
-        });
-    });
 }
 
 export function fetchCharts(schemaQuery: SchemaQuery, containerPath?: string): Promise<List<DataViewInfo>> {
