+++ packages/components/src/internal/components/editable/Cell.spec.tsx
@@ -18,7 +18,7 @@ import { mount, ReactWrapper } from 'enzyme';
 
 import { List } from 'immutable';
 
-import { ValueDescriptor } from '../../models';
+import { ValueDescriptor } from './models';
 
 import { CELL_SELECTION_HANDLE_CLASSNAME } from '../../constants';
