+++ packages/components/src/internal/components/editable/models.spec.ts
@@ -15,12 +15,12 @@
  */
 import { fromJS, List, Map, Set } from 'immutable';
 
-import sampleSet2QueryInfo from '../test/data/sampleSet2-getQueryDetails.json';
+import sampleSet2QueryInfo from '../../../test/data/sampleSet2-getQueryDetails.json';
 
-import { SchemaQuery } from '../public/SchemaQuery';
-import { QueryColumn } from '../public/QueryColumn';
-import { QueryInfo } from '../public/QueryInfo';
-import { makeTestQueryModel } from '../public/QueryModel/testUtils';
+import { SchemaQuery } from '../../../public/SchemaQuery';
+import { QueryColumn } from '../../../public/QueryColumn';
+import { QueryInfo } from '../../../public/QueryInfo';
+import { makeTestQueryModel } from '../../../public/QueryModel/testUtils';
 
 import { CellMessage, EditorModel, getPkData, ValueDescriptor } from './models';
