+++ packages/components/src/internal/components/editable/utils.spec.ts
@@ -2,7 +2,7 @@ import { List, OrderedMap } from 'immutable';
 
 import { QueryModel } from '../../../public/QueryModel/QueryModel';
 import { LoadingState } from '../../../public/LoadingState';
-import { EditorModel } from '../../models';
+import { EditorModel } from './models';
 import { ASSAY_WIZARD_MODEL } from '../../../test/data/constants';
 
 import { getEditorTableData } from './utils';
